split forwards and backwards mappers
